Fix documentation for custom beat creation
When I tried these instructions, they did not work. When running `mage GenerateCustomBeat` from the newly created directory in ` ${GOPATH}/src/github.com/{user}` an error is thrown: `No .go files marked with the mage build tag in this directory.`. The process works correctly when the command is run from the Beats directory.